backend/requirements/base.txt
# requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
minio==7.1.16
kafka-python==2.0.2
apscheduler==3.10.4
python-jose==3.3.0
python-multipart==0.0.6
strawberry-graphql[fastapi]==0.215.1
pydantic-settings==2.1.0
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
weaviate-client==3.26.2
neo4j==5.13.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6